Extract Cli::Commands::Base as a helper module
